let exifr read the file insteadof the app guessing the size.  #277

+++ test/folder-reset.ts
@@ -8,6 +8,9 @@ const dir = process.argv[2];
 if(dir === 'test'){
   throw new Error('test folder is not allowed to be deleted');
 }
+if(dir === '--exclude'){
+  throw new Error('--exclude folder is not allowed to be deleted/created');
+}
 if (fs.existsSync(dir)) {
   console.log('deleting folder:' + dir);
   fs.rmSync(dir, {recursive: true});
